requirements.txt
fastapi
pymongo==4.7
uvicorn
python-dotenv
passlib[bcrypt]